nova_server/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
python-dotenv
